Flight-Software/requirements.txt
pyftdi
pyusb
pyserial
pytest
psutil==5.8.0
numpy==1.17.3
Adafruit_BBIO==1.2.0
brotli==1.0.9
chardet==4.0.0
cryptography==3.4.7
Cython==0.29.24
mock==4.0.3
Pillow==8.3.1
protobuf==3.17.3
pyOpenSSL==20.0.1
simplejson==3.17.3
unicodedata2==13.0.0.post2

